+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `approverEditability` on form elements so they can be edited by approvers
 - optional `s3ObjectVersionId` to `Forms.getSubmissionData()`
 - `formSubmissionMetaEdits`, `formSubmissionSchedulingBooking` and `taskCompletion` to `Forms.getFormSubmissionMeta()` result
+- validation for min and max years for MySQL DATETIME columns to form properties: `publishStartDate` and `publishEndDate`
 
 ## [13.2.3] - 2026-08-26